requirements.txt
Flask
azure-storage-blob
azure-cosmos
gunicorn  # if using for deployment (optional for local development)
